bug fixes for targettype and related to: - new_type_argument - class_literal_generic_or_array

+++ src/share/classes/com/sun/tools/javac/comp/TargetType.java
@@ -220,7 +220,7 @@ public enum TargetType {
         if (targets == null)
             targets = buildTargets();
 
-        if (tag == UNKNOWN.targetTypeValue)
+        if (((byte)tag) == ((byte)UNKNOWN.targetTypeValue))
             return UNKNOWN;
         // we can optimize the algorithm a bit: binary search?
         if (tag < 0 || tag >= targets.length)
